Replot graphs empty before the actual update
There appears to be an issue with plotly where sequential updates that
simply append to data can have points of the old data not correctly
removed from the plot.  This results in "phantom" traces that are
especially visible when the plot changes axis scaling.  To work around
this, do a plot update with no data before applying the actual plot
update.