Make archive server less synchronous
Make the archive server less synchronous with respect to IO operation.
This moves the long-running operations (transaction commit and file
list) to a separate worker thread, so they don't block the asyncio
thread handling connections.  It also changes the protocol so that
IO operations respond immediately, so that the client doesn't time
out waiting for the response while the server works.